# perf: use mimalloc