### Before

```
Running "satisfies" suite...
Progress: 100%

  Native:
    739 151 ops/s, ±0.30%   | fastest

  JavaScript :
    189 590 ops/s, ±0.12%   | slowest, 74.35% slower

Finished 2 cases!
  Fastest: Native
  Slowest: JavaScript
```

### After

```
Running "satisfies" suite...
Progress: 100%

  Native:
    1 200 843 ops/s, ±0.12%   | fastest

  JavaScript :
    188 835 ops/s, ±0.26%     | slowest, 84.27% slower

Finished 2 cases!
  Fastest: Native
  Slowest: JavaScript 
```